Role checking/Date verification/Dictionary Dumping
Added the ability to dump the schedule however you must have the role to do so. Role checking is now enabled for remove/add as only team captains should be allowed. Fixed some date checking and time checking as you were not allowed to schedule for anything but whole numbers and the time would exceed 11pm.